\name{alphabets-package}
\alias{alphabets-package}
\alias{alphabets}
\docType{package}
\title{
\packageTitle{alphabets}
}
\description{
\packageDescription{alphabets}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{alphabets}
\packageIndices{alphabets}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{alphabets}

Maintainer: \packageMaintainer{alphabets}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
